rag-app/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
python-docx
groq
python-multipart